add allowance key when creating an asset
