+++ packages/schema/src/roles/allowances.ts
@@ -1,0 +1,7 @@
+export enum AllowanceInterval {
+  Daily = 'daily',
+  Weekly = 'weekly',
+  Monthly = 'monthly',
+  Quarterly = 'quarterly',
+  Yearly = 'yearly',
+}
